# backend/bdd/doc/openapi/bddOpenApi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Generated API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5001"
  } ],
  "paths" : {
...
    "/api/database/groupOrders/order/{orderId}" : {
      "get" : {
        "summary" : "getGroupOrderByOrderId",
        "description" : "Generated from method getGroupOrderByOrderId",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "serialVersionUID" : {
                "type" : "integer"
              },
              "mostSigBits" : {
                "type" : "integer"
              },
              "leastSigBits" : {
                "type" : "integer"
              },
              "jla" : {
                "type" : "object"
              },
              "NIBBLES" : {
                "type" : "object"
              },
              "$assertionsDisabled" : {
                "type" : "boolean"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/database/orders/individual/create" : {
      "post" : {
        "summary" : "createIndividualOrder",
        "description" : "Generated from method createIndividualOrder",
        "parameters" : [ {
          "name" : "IndividualOrderDTO",
          "in" : "body",
          "schema" : {
            "$ref" : "#/components/schemas/IndividualOrderDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/database/orders/individual/{id}" : {
      "get" : {
        "summary" : "getIndividualOrderById",
        "description" : "Generated from method getIndividualOrderById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "serialVersionUID" : {
                "type" : "integer"
              },
              "mostSigBits" : {
                "type" : "integer"
              },
              "leastSigBits" : {
                "type" : "integer"
              },
              "jla" : {
                "type" : "object"
              },
              "NIBBLES" : {
                "type" : "object"
              },
              "$assertionsDisabled" : {
                "type" : "boolean"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/database/orders/create" : {
      "post" : {
        "summary" : "createOrder",
        "description" : "Generated from method createOrder",
        "parameters" : [ {
          "name" : "OrderDTO",
          "in" : "body",
          "schema" : {
            "$ref" : "#/components/schemas/OrderDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/database/orders/{id}" : {
      "delete" : {
        "summary" : "deleteOrder",
        "description" : "Generated from method deleteOrder",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "serialVersionUID" : {
                "type" : "integer"
              },
              "mostSigBits" : {
                "type" : "integer"
              },
              "leastSigBits" : {
                "type" : "integer"
              },
              "jla" : {
                "type" : "object"
              },
              "NIBBLES" : {
                "type" : "object"
              },
              "$assertionsDisabled" : {
                "type" : "boolean"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      },
      "get" : {
        "summary" : "getOrderById",
        "description" : "Generated from method getOrderById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "object",
            "properties" : {
              "serialVersionUID" : {
                "type" : "integer"
              },
              "mostSigBits" : {
                "type" : "integer"
              },
              "leastSigBits" : {
                "type" : "integer"
              },
              "jla" : {
                "type" : "object"
              },
              "NIBBLES" : {
                "type" : "object"
              },
              "$assertionsDisabled" : {
                "type" : "boolean"
              }
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/database/orders/individual/update" : {
      "put" : {
        "summary" : "updateOrder",
        "description" : "Generated from method updateOrder",
        "parameters" : [ {
          "name" : "IndividualOrderDTO",
          "in" : "body",
          "schema" : {
            "$ref" : "#/components/schemas/IndividualOrderDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/database/orders/update" : {
      "put" : {
        "summary" : "updateOrder",
        "description" : "Generated from method updateOrder",
        "parameters" : [ {
          "name" : "OrderDTO",
          "in" : "body",
          "schema" : {
            "$ref" : "#/components/schemas/OrderDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HttpResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "HttpResponse" : {
        "type" : "object",
        "properties" : {
          "statusCode" : {
            "type" : "integer"
          },
          "body" : {
            "type" : "string"
          }
        }
      },
...
      "IndividualOrderDTO" : {
        "type" : "object",
        "properties" : {
          "deliveryDetails" : {
            "type" : "object"
          }
        }
      },
      "OrderDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "restaurantId" : {
            "type" : "string"
          },
          "userId" : {
            "type" : "string"
          },
          "items" : {
            "type" : "array"
          },
          "status" : {
            "type" : "string"
          },
          "slot" : {
            "type" : "object"
          }
        }
      },
...
    }
  }
}